backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.10.1
email-validator==2.2.0
greenlet==3.2.4
bcrypt==4.0.1
